5 April 1:06AM
Completed:
1. Checkout process successful, stripe checkout url received.
2. Stripe webhook listening
forwarded to container, although require manual scripting

TO-DO:
1. Configure stripe event handling
2. Kafka still not working

+++ Backend/PlaceAnOrderOrchestrator/src/main/java/com/esd/PlaceAnOrderOrchestrator/dto/OrderStatusUpdateRequest.java
@@ -4,6 +4,6 @@ import lombok.Data;
 
 @Data
 public class OrderStatusUpdateRequest {
-    private Long order_id;
+    private Integer order_id;
     private String new_order_status;
 }
