+++ Backend/PlaceAnOrderOrchestrator/src/main/java/com/esd/PlaceAnOrderOrchestrator/dto/StockUpdateRequest.java
@@ -4,6 +4,6 @@ import lombok.Data;
 
 @Data
 public class StockUpdateRequest {
-    private Long item_id;
+    private Integer item_id;
     private Integer item_quantity;   
 }
